requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
scipy>=1.11.0
matplotlib>=3.7.0
Pillow>=10.0.0
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0

# Deep learning models
transformers>=4.35.0
timm>=0.9.0

# Optional: SAM (Segment Anything Model)
# Install from: pip install git+https://github.com/facebookresearch/segment-anything.git
# Download checkpoint: https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth

# Optional: CLIP
# Install from: pip install git+https://github.com/openai/CLIP.git

# Optional: LLaVA
accelerate>=0.24.0
